+++ .github/workflows/close_jira_issue_by_pr_merge.yml
@@ -37,5 +37,5 @@ jobs:
           INSTANCE_URL: ${{ secrets.GOLD_SERVER_URL_NG }}
         run: |
           echo "Checking for related Jira issues to PR: $PR_NUMBER"
-          cd Utils/github_workflow_scripts/jira_integration_scripts
+          cd .github/github_workflow_scripts/jira_integration_scripts
           poetry run python ./link_pr_to_jira_issue.py --url "$INSTANCE_URL" --pr_num "$PR_NUMBER" --pr_link "$PR_LINK" --pr_title "$PR_TITLE" --pr_body "$PR_BODY" --is_merged --username $USERNAME --password $PASSWORD
