+++ .github/workflows/link_edited_pr_to_jira_issue.yml
@@ -37,6 +37,6 @@ jobs:
           INSTANCE_URL: ${{ secrets.GOLD_SERVER_URL_NG }}
         run: |
           echo "Checking for related Jira issues to PR: $PR_NUMBER"
-          cd Utils/github_workflow_scripts/jira_integration_scripts
+          cd .github/github_workflow_scripts/jira_integration_scripts
           echo --pr_num $PR_NUMBER --pr_link $PR_LINK --pr_title $PR_TITLE --pr_body $PR_BODY --no-is_merged
           poetry run python ./link_pr_to_jira_issue.py --url "$INSTANCE_URL" --pr_num "$PR_NUMBER" --pr_link "$PR_LINK" --pr_title "$PR_TITLE" --pr_body "$PR_BODY" --no-is_merged --username $USERNAME --password $PASSWORD
